% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_mass.R
\name{calc_mass}
\alias{calc_mass}
\title{Calculate the mass of a substance containing a specified number of moles.}
\usage{
calc_mass(substance, val, u = "mol")
}
\arguments{
\item{substance}{Enter the common name of the chemical compound or substance.}

\item{val}{The number of moles.}

\item{u}{The unit, select from "mol", "umol", "pmol", "fmol" with default set to moles.}
}
\value{
The mass of the substance containing the specified number of moles, in grams.
}
\description{
Calculates the mass of a specified substance in grams that will contain the specified number of moles.
}
\examples{
"ammonium"

}
