% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_mol.R
\name{calc_mol}
\alias{calc_mol}
\title{Calculate the number of moles in a specified amount of a chemical substance.}
\usage{
calc_mol(substance, mass, u = "g")
}
\arguments{
\item{substance}{Enter the common name of the chemical compound or substance.}

\item{mass}{The mass.}

\item{u}{The unit, select from "g", "ug", "pg", "fg" with default set to grams}
}
\value{
The number of moles of substance in the specified mass.
}
\description{
Calculates the number of moles in a specified amount of a chemical substance.
}
\examples{
"ammonium"

}
