% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_antibody_dilutions.R
\name{find_dilutions_thermo}
\alias{find_dilutions_thermo}
\title{Retrieve the recommended dilutions per application type for a given antibody.}
\usage{
find_dilutions_thermo(catalogue_number, type)
}
\arguments{
\item{catalogue_number}{Catalogue number of the antibody, entered as a number.}

\item{type}{Type of antibody (primary or secondary).}
}
\value{
Dataframe with listed applications tested by the vendor, and concentrations for each.
}
\description{
Retrieves recommended dilutions for a given antibody from the vendor for all tested application types.
}
\examples{
31594

}
